Prediction Model/requirements.txt
# Core libraries
numpy>=1.21.0
pandas>=1.3.0

# Data fetching and analysis
yfinance>=0.1.67
requests>=2.25
pandas_market_calendars>=4.6.1
pytorch-forecasting>=1.3.0

# Visualization
matplotlib>=3.3
seaborn>=0.11
torch>=2.1.0  # for SummaryWriter
pydot>=1.4.2
graphviz>=0.20.1
streamlit>=1.42.1
plotly>=4.14.3
streamlit-autorefresh>=1.0.1

# Deep learning (TensorFlow with GPU support)
tensorflow>=2.18

# Machine Learning utilities
scikeras>=0.12.0
scikit-learn>=0.24.2
statsmodels>=0.13
scipy>=1.5
tqdm>=4.60
scikit-image>=0.18

# Time-series feature engineering
PyWavelets>=1.1.1
hmmlearn>=0.2.7

# Additional libraries
pyyaml>=5.4
optuna>=2.10.0
xgboost>=1.5.0
